Removed some more Win64 warnings. On linux all tests pass. Awaiting the next round of warnings on the regression page.....
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8510 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/aterm/include/aterm/aterm2.h
@@ -254,7 +254,7 @@ int ATcompare(ATerm t1, ATerm t2);
 void ATsetChecking(ATbool on);
 ATbool ATgetChecking(void);
 
-extern int at_gc_count;
+extern size_t at_gc_count;
 #define ATgetGCCount()    (at_gc_count)
 
 size_t  ATcalcTextSize(ATerm t);
